Refactor BaseJavaModule and ReactContextBaseJavaModule (#39823)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39823

In this diff I'm refactoring BaseJavaModule and ReactContextBaseJavaModule to simplify class hierarchy.

ReactContextBaseJavaModule will be deprecated in the new architecture

bypass-github-export-checks

Reviewed By: cortinico

Differential Revision: D49930340

fbshipit-source-id: c56fda685382a8f5cf549533687c9c9bb843316b
