+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/bridge/BaseJavaModule.java
@@ -7,7 +7,14 @@
 
 package com.facebook.react.bridge;
 
+import static com.facebook.infer.annotation.ThreadConfined.ANY;
+
 import androidx.annotation.Nullable;
+import com.facebook.common.logging.FLog;
+import com.facebook.infer.annotation.Assertions;
+import com.facebook.infer.annotation.ThreadConfined;
+import com.facebook.react.common.ReactConstants;
+import com.facebook.react.common.build.ReactBuildConfig;
 import java.util.Map;
 
 /**
@@ -44,6 +51,16 @@ public abstract class BaseJavaModule implements NativeModule {
   public static final String METHOD_TYPE_PROMISE = "promise";
   public static final String METHOD_TYPE_SYNC = "sync";
 
+  private final @Nullable ReactApplicationContext mReactApplicationContext;
+
+  public BaseJavaModule() {
+    this(null);
+  }
+
+  public BaseJavaModule(@Nullable ReactApplicationContext reactContext) {
+    mReactApplicationContext = reactContext;
+  }
+
   /** @return a map of constants this module exports to JS. Supports JSON types. */
   public @Nullable Map<String, Object> getConstants() {
     return null;
@@ -69,5 +86,41 @@ public abstract class BaseJavaModule implements NativeModule {
   @Override
   public void invalidate() {
     onCatalystInstanceDestroy();
+  }
+
+  /** Subclasses can use this method to access catalyst context passed as a constructor. */
+  protected final ReactApplicationContext getReactApplicationContext() {
+    return Assertions.assertNotNull(
+        mReactApplicationContext,
+        "Tried to get ReactApplicationContext even though NativeModule wasn't instantiated with one");
+  }
+
+  /**
+   * Subclasses can use this method to access catalyst context passed as a constructor. Use this
+   * version to check that the underlying CatalystInstance is active before returning, and
+   * automatically have SoftExceptions or debug information logged for you. Consider using this
+   * whenever calling ReactApplicationContext methods that require the Catalyst instance be alive.
+   *
+   * <p>This can return null at any time, but especially during teardown methods it's
+   * possible/likely.
+   *
+   * <p>Threading implications have not been analyzed fully yet, so assume this method is not
+   * thread-safe.
+   */
+  @ThreadConfined(ANY)
+  protected @Nullable final ReactApplicationContext getReactApplicationContextIfActiveOrWarn() {
+    if (mReactApplicationContext.hasActiveReactInstance()) {
+      return mReactApplicationContext;
+    }
+
+    // We want to collect data about how often this happens, but SoftExceptions will cause a crash
+    // in debug mode, which isn't usually desirable.
+    String msg = "React Native Instance has already disappeared: requested by " + getName();
+    if (ReactBuildConfig.DEBUG) {
+      FLog.w(ReactConstants.TAG, msg);
+    } else {
+      ReactSoftExceptionLogger.logSoftException(ReactConstants.TAG, new RuntimeException(msg));
+    }
+    return null;
   }
 }
